Started initial implementation for packing model
